# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/actas-inicio": {
      "get": {
        "operationId": "ActaInicioController_findAll",
        "parameters": [
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Campos a incluir en la respuesta (separados por comas)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Dirección del ordenamiento (ASC o DESC)",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Número máximo de registros a retornar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Número de registros a saltar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Filtros en formato JSON. Ejemplo: {\"contrato_general_id\":1,\"activo\":true}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Relaciones a incluir (separadas por comas). Ejemplo: contrato_general",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de actas de inicio",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActaInicio"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener todas las actas de inicio",
        "tags": [
          "actas-inicio"
        ]
      },
      "post": {
        "operationId": "ActaInicioController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearActaInicioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Acta de inicio creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActaInicio"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "404": {
            "description": "Contrato general no encontrado"
          }
        },
        "summary": "Crear una nueva acta de inicio",
        "tags": [
          "actas-inicio"
        ]
      }
    },
    "/actas-inicio/{id}": {
      "get": {
        "operationId": "ActaInicioController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del acta de inicio",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acta de inicio encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActaInicio"
                }
              }
            }
          },
          "404": {
            "description": "Acta de inicio no encontrada"
          }
        },
        "summary": "Obtener una acta de inicio por ID",
        "tags": [
          "actas-inicio"
        ]
      },
      "put": {
        "operationId": "ActaInicioController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del acta de inicio a actualizar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarActaInicioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acta de inicio actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActaInicio"
                }
              }
            }
          },
          "404": {
            "description": "Acta de inicio no encontrada"
          }
        },
        "summary": "Actualizar una acta de inicio",
        "tags": [
          "actas-inicio"
        ]
      },
      "delete": {
        "operationId": "ActaInicioController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del acta de inicio a eliminar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acta de inicio eliminada"
          },
          "404": {
            "description": "Acta de inicio no encontrada"
          }
        },
        "summary": "Eliminar lógicamente una acta de inicio",
        "tags": [
          "actas-inicio"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gestion Contractual CRUD",
    "description": "API para la gestión contractual",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ActaInicio": {
        "type": "object",
        "properties": {}
      },
      "CrearActaInicioDto": {
        "type": "object",
        "properties": {
          "usuario_id": {
            "type": "number",
            "example": 1,
            "description": "Id del usuario que creó el acta"
          },
          "usuario_legado": {
            "type": "string",
            "example": "usuario_legado",
            "description": "Usuario heredado del sistema anterior"
          },
          "descripcion": {
            "type": "string",
            "example": "Descripción del acta",
            "description": "Descripción breve del acta"
          },
          "fecha_inicio": {
            "type": "string",
            "example": "2023-10-24",
            "description": "Fecha de inicio del acta (formato ISO)"
          },
          "fecha_fin": {
            "type": "string",
            "example": "2023-11-24",
            "description": "Fecha de finalización del acta (formato ISO)"
          },
          "contrato_general_id": {
            "type": "number",
            "example": 1,
            "description": "Id del contrato general relacionado"
          },
          "activo": {
            "type": "boolean",
            "example": true,
            "description": "Indica si el acta está activa"
          }
        },
        "required": [
          "usuario_id",
          "usuario_legado",
          "descripcion",
          "fecha_inicio",
          "fecha_fin",
          "contrato_general_id",
          "activo"
        ]
      },
      "ActualizarActaInicioDto": {
        "type": "object",
        "properties": {
          "usuario_id": {
            "type": "number",
            "example": 1,
            "description": "Id del usuario que creó el acta"
          },
          "usuario_legado": {
            "type": "string",
            "example": "usuario_legado",
            "description": "Usuario heredado del sistema anterior"
          },
          "descripcion": {
            "type": "string",
            "example": "Descripción del acta",
            "description": "Descripción breve del acta"
          },
          "fecha_inicio": {
            "type": "string",
            "example": "2023-10-24",
            "description": "Fecha de inicio del acta (formato ISO)"
          },
          "fecha_fin": {
            "type": "string",
            "example": "2023-11-24",
            "description": "Fecha de finalización del acta (formato ISO)"
          },
          "contrato_general_id": {
            "type": "number",
            "example": 1,
            "description": "Id del contrato general relacionado"
          },
          "activo": {
            "type": "boolean",
            "example": true,
            "description": "Indica si el acta está activa"
          }
        }
      },
...
    }
  }
}